debug: option to print using mnl format
nftables uses the libmnl library that has a particular debugging format.
This format is also used for nftables --debug=all.
google/nftables golang library to interact with nftables uses this
library to interface with netlink directly.
Since these libraries are not well documented, is common to have to dump
the netlink bits to compare.

This commit implements an option to choose the same format used in
nftables by adding a new option to the NLDEBUG environmenta variable and
using the format=mnt key value.

Change-Id: Ifcddb31d1561074942ad8d6af74b134540eca5fe
Signed-off-by: Antonio Ojea <aojea@google.com>