debug: option to print using mnl format
nftables uses the libmnl library that has a particular debugging format.
This format is also used for nftables --debug=all.
google/nftables golang library to interact with nftables uses this
library to interface with netlink directly.
Since these libraries are not well documented, is common to have to dump
the netlink bits to compare.

This commit implements an option to choose the same format used in
nftables by adding a new option to the NLDEBUG environmenta variable and
using the format=mnl key value.

For Linux systems mnl is the default format.

Change-Id: I5f63753b3ada4d25565fe0bddac35e010510a34f