Add Sequence field to OpError for better error correlation (#236)
When a message with an errno is returned, it triggers an error in
checkMessage and is not passed back to the caller. This can make it
difficult to match the error to the specific message that caused it
when using conn.SendMessages.

This change includes the message’s Sequence value in OpError, making
it easier to identify which message the error corresponds to.